Moved typing.py to the main geomloss folder.
