+++ geomloss/ot/abstract_solvers/sinkhorn_ot.py
@@ -1,5 +1,5 @@
 from ... import backends as bk
-from ..typing import (
+from ...typing import (
     List,
     RealTensor,
     Optional,
